Move open path (polyline) to the line family: line;path (#121)
A polyline is a 1-D route — the multi-segment generalization of
line;connect — not a 2-D polygon. Move it from the polygon family to
the line family:

- enum PolygonConstructions.path → LineConstructions.path (114)
- PolylineConstruction → LineConstructions.ts (registered in lineConstructions)
- src/elements/polygon/PolylineElement.ts → src/elements/line/PolylineElement.ts
- params/demo/docs: polygon;path → line;path (E.Polygon.path → E.Line.path)
- path tests moved PolygonTest → LineTest (test-guided)

PolylineElement still extends GeomElement (not LineElement), so it is
never subject to the line-endpoint expansion contract. 285 unit + 700
snapshot green.
